Turn on tests again. (#26)
This enables tests again on PRs. See the .github/workflows/flutter_packages.yaml file for details.

They got broken because I set it up with specific paths. This should now search for directories with pubspec.yaml files and run the tests in each one.

I also fixed all the existing issues, and normalized the SDK constraints.

It does do the setup for Flutter for each package, but the Flutter download is cached. I tried making it all one step, but then you can't see separate steps in the Github UI, and they can't run in parallel.