Add unit tests for `fluter_genui` package (#38)
Adds a bunch of unit tests for flutter_genui so we can make changes with confidence. It also introduces a GenerativeModelInterface class that abstracts the generative model somewhat, although the client is still Gemini-specific. It's just to make writing tests easier because Firebase AI Logic (unlike google_generative_ai) has no testing interface for GenerationConfig, and it's a final class, so it can't be mocked.