Fix the infinite loop, add text response. (#191)
This fixes the infinite loop (again!) by correctly parsing the return value from the inference.  It was trying to get the "parameters" field, which was never set, and so was always null, causing an infinite tool loop.

I also added an optional text response in the travel app because if the user asks for an impossible destination ("Plan a trip to hell"), then it can return "false", and needs a way to communicate why.  It also makes it a lot more conversational even when it succeeds and includes a message.