chore(simple_chat): fix instruction formatting (#326)
Fixes a grammatical typo in the prompt of the simple_chat example app. Also reformats part of the instruction as a single line, avoiding hard line wraps and indentations in the middle of sentences.

This might have negligible effect on LLM output, but I figure this is a very cheap improvement to make.